# sdks/sdkwork-im-backend-sdk/openapi/sdkwork-im-backend-api.openapi.yaml
openapi: 3.1.2
info:
  title: Sdkwork IM Backend Management API
  version: 0.1.0
  description: Backend management OpenAPI contract for operator, governance, control, and admin APIs under one SDK family.
servers:
  - url: http://127.0.0.1:18079
security:
  - AuthToken: []
    AccessToken: []
...
paths:
...
  /backend/v3/api/control/social/friend_requests/{requestId}/decline:
    post:
      operationId: social.friendRequests.decline
      parameters:
        - description: Aggregate identifier.
          in: path
          name: requestId
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeclineFriendRequestRequest'
        description: Social action payload.
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SdkWorkApiResponse'
                  - type: object
                    required:
                      - data
                    properties:
                      data:
                        type: object
                        required:
                          - item
                        properties:
                          item:
                            $ref: '#/components/schemas/SocialFriendRequestCommitResponse'
          description: Successful response.
        '400':
          description: Invalid request.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Authentication required.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Permission denied.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: Resource not found.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '409':
          description: Request conflicts with current control-plane state.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '503':
          description: Control-plane dependency is unavailable.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
      security:
        - AuthToken: []
          AccessToken: []
      summary: Decline a friend request.
      tags:
        - control
      x-sdkwork-owner: sdkwork-im
      x-sdkwork-api-authority: sdkwork-im.backend
      x-sdkwork-request-context: WebRequestContext
      x-sdkwork-api-surface: backend-api
...
components:
  securitySchemes:
    AuthToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
    AccessToken:
      type: apiKey
      in: header
      name: Access-Token
...
  schemas:
    ProblemDetail:
      type: object
      additionalProperties: true
      required:
        - type
        - title
        - status
        - code
        - traceId
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 100
          maximum: 599
        detail:
          type: string
        instance:
          type: string
        code:
          $ref: '#/components/schemas/SdkWorkPlatformErrorCode'
        traceId:
          type: string
          format: uuid
          description: Server-owned request correlation id.
        i18nKey:
          type: string
          description: Optional stable localization key such as errors.result.40001.
        locale:
          type: string
          description: Optional effective BCP 47 locale used by framework message mapping.
        errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
...
    DeclineFriendRequestRequest:
      additionalProperties: false
      properties:
        declinedAt:
          type: string
        declinedByUserId:
          type: string
        eventId:
          type: string
      required:
        - eventId
        - declinedByUserId
        - declinedAt
      type: object
...
    SocialFriendRequestCommitResponse:
      additionalProperties: true
      description: Friend request write result plus persistence metadata.
      properties: {}
      type: object
...
    SdkWorkApiResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - data
        - traceId
      properties:
        code:
          type: integer
          format: int32
          enum:
            - 0
          default: 0
          minimum: 0
          maximum: 0
        data:
          description: Operation-specific payload typed per response schema.
        traceId:
          type: string
          format: uuid
          description: Server-owned request correlation id.
...
    SdkWorkPlatformErrorCode:
      type: integer
      format: int32
      minimum: 40001
      maximum: 79999
      description: Platform or domain error code per API_SPEC.md §15.3.
    FieldError:
      type: object
      additionalProperties: false
      required:
        - field
        - message
      properties:
        field:
          type: string
        message:
          type: string
        code:
          type: integer
          format: int32
          minimum: 40011
          maximum: 40099
        i18nKey:
          type: string
        params:
          type: object
          additionalProperties:
            oneOf:
              - type: string
              - type: number
              - type: integer
              - type: boolean
...